backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0

pipecat-ai>=0.0.47

numpy>=1.26.0
scipy>=1.11.0
pyaudio>=0.2.11

google-generativeai>=0.8.0
google-cloud-speech>=2.20.0
google-cloud-texttospeech>=2.16.0

onnxruntime>=1.16.0

deepgram-sdk>=3.2.0

anyio>=3.7.1,<4.0.0
protobuf>=3.19.5,<5.0.0dev
aiohttp>=3.8.0
